add centralized provider type registry
