+++ packages/sdk/src/lend/namespaces/BaseLendNamespace.ts
@@ -1,7 +1,6 @@
-import type { LendProvider } from '@/lend/core/LendProvider.js'
 import type { AaveLendProvider } from '@/lend/providers/aave/AaveLendProvider.js'
 import type { MorphoLendProvider } from '@/lend/providers/morpho/MorphoLendProvider.js'
-import type { LendProviderConfig } from '@/types/actions.js'
+import type { LendProviders } from '@/types/providers.js'
 import type {
   GetLendMarketParams,
   GetLendMarketsParams,
@@ -9,10 +8,7 @@ import type {
   LendMarketId,
 } from '@/types/lend/index.js'
 
-export type LendProviders = {
-  morpho?: LendProvider<LendProviderConfig>
-  aave?: LendProvider<LendProviderConfig>
-}
+export type { LendProviders } from '@/types/providers.js'
 
 /**
  * Base Lend Namespace
