+++ packages/sdk/src/types/swap/base.ts
@@ -1,7 +1,6 @@
 import type { Address } from 'viem'
 
 import type { SupportedChainId } from '@/constants/supportedChains.js'
-import type { SwapProvider } from '@/swap/core/SwapProvider.js'
 import type { SwapProviderName } from '@/types/actions.js'
 import type { Asset } from '@/types/asset.js'
 import type { TransactionData } from '@/types/transaction.js'
@@ -13,14 +12,7 @@ import type {
 export { SwapProvider } from '@/swap/core/SwapProvider.js'
 export { ActionsSwapNamespace } from '@/swap/namespaces/ActionsSwapNamespace.js'
 export { WalletSwapNamespace } from '@/swap/namespaces/WalletSwapNamespace.js'
-
-/**
- * Map of available swap providers keyed by provider name
- */
-export type SwapProviders = {
-  uniswap?: SwapProvider<SwapProviderConfig>
-  velodrome?: SwapProvider<SwapProviderConfig>
-}
+export type { SwapProviders } from '@/types/providers.js'
 
 /**
  * Swap provider configuration
